+++ deployment/updates/scripts/add_permissions/2020_07_26_add_permission_widget_object_privileges_field.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package deployment
+ * Add permissions to privileges field on widget object
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/object.KalturaWidget.ini';
+passthru("php $script $config");