+++ infra/plugins/interfaces/IKalturaImageTransformationExecutor.php
@@ -8,5 +8,5 @@
 interface IKalturaImageTransformationExecutor extends IKalturaBase
 {
 	public function getImageFile($entry, $version, $width, $height, $type, $bgcolor, $quality, $src_x, $src_y, $src_w, $src_h,
-							 $vid_sec, $vid_slice, $vid_slices, $orig_image_path, $density, $stripProfiles, $format, $start_sec, $end_sec);
+							 $vid_sec, $vid_slice, $vid_slices, $orig_image_path, $density, $stripProfiles, $format, $fileSync, $start_sec, $end_sec);
 }